Changed top earner ID to a long, updated mocked response to a long

+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/network/rest/wpcom/wc/orderstats/TopEarnersStatsApiResponse.kt
@@ -10,7 +10,7 @@ class TopEarnersStatsApiResponse : Response {
 
     class TopEarner {
         @SerializedName("ID")
-        val id: Int? = 0
+        val id: Long? = 0
         val currency: String? = ""
         val image: String? = ""
         val name: String? = ""
